+++ backend/springboot/src/main/java/com/workmaite/domain/chat/entity/ChatMessage.java
@@ -1,10 +1,9 @@
 package com.workmaite.domain.chat.entity;
 
 import jakarta.persistence.*;
+import java.time.LocalDateTime;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "chat_messages")
@@ -14,38 +13,38 @@ import java.time.LocalDateTime;
 @Builder
 public class ChatMessage {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Integer id;
 
-    @Column(name = "thread_id", length = 100, nullable = false)
-    private String threadId;
+  @Column(name = "thread_id", length = 100, nullable = false)
+  private String threadId;
 
-    @Column(name = "user_id", nullable = false)
-    private Long userId;
+  @Column(name = "user_id", nullable = false)
+  private Integer userId;
 
-    @Column(length = 10, nullable = false)
-    private String role;
+  @Column(length = 10, nullable = false)
+  private String role;
 
-    @Column(columnDefinition = "TEXT", nullable = false)
-    private String content;
+  @Column(columnDefinition = "TEXT", nullable = false)
+  private String content;
 
-    @Column(name = "file_path", length = 500)
-    private String filePath;
+  @Column(name = "file_path", length = 500)
+  private String filePath;
 
-    @Column(name = "file_name", length = 255)
-    private String fileName;
+  @Column(name = "file_name", length = 255)
+  private String fileName;
 
-    @Column(name = "context_type", length = 20)
-    private String contextType;
+  @Column(name = "context_type", length = 20)
+  private String contextType;
 
-    @Column(name = "meeting_id")
-    private Long meetingId;
+  @Column(name = "meeting_id")
+  private Integer meetingId;
 
-    @Column(name = "session_id")
-    private Long sessionId;
+  @Column(name = "session_id")
+  private Integer sessionId;
 
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
+  @CreationTimestamp
+  @Column(name = "created_at", updatable = false)
+  private LocalDateTime createdAt;
 }
