+++ backend/springboot/src/main/java/com/workmaite/domain/company/entity/Company.java
@@ -1,18 +1,14 @@
 package com.workmaite.domain.company.entity;
 
 import jakarta.persistence.*;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
-/**
- * 회사(테넌트) 엔티티 — companies 테이블 (P1-7② 정규화, MT-4).
- * users.company 자유 텍스트를 대체하는 정규화된 회사 엔티티.
- */
+/** 회사(테넌트) 엔티티 — companies 테이블 (P1-7② 정규화, MT-4). users.company 자유 텍스트를 대체하는 정규화된 회사 엔티티. */
 @Entity
 @Table(name = "companies")
 @Getter
@@ -21,18 +17,18 @@ import java.time.LocalDateTime;
 @Builder
 public class Company {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Integer id;
 
-    @Column(nullable = false, unique = true, length = 100)
-    private String name;
+  @Column(nullable = false, unique = true, length = 100)
+  private String name;
 
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
+  @Column(name = "created_at", nullable = false, updatable = false)
+  private LocalDateTime createdAt;
 
-    @PrePersist
-    protected void onCreate() {
-        createdAt = LocalDateTime.now();
-    }
+  @PrePersist
+  protected void onCreate() {
+    createdAt = LocalDateTime.now();
+  }
 }
