+++ backend/springboot/src/main/java/com/workmaite/domain/meetings/repository/MeetingMemberRepository.java
@@ -2,24 +2,23 @@ package com.workmaite.domain.meetings.repository;
 
 import com.workmaite.domain.meetings.entity.MeetingMember;
 import com.workmaite.domain.meetings.entity.MeetingMemberRole;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MeetingMemberRepository extends JpaRepository<MeetingMember, Long> {
+public interface MeetingMemberRepository extends JpaRepository<MeetingMember, Integer> {
 
-    List<MeetingMember> findByMeetingId(Long meetingId);
+  List<MeetingMember> findByMeetingId(Integer meetingId);
 
-    Optional<MeetingMember> findByMeetingIdAndUserId(Long meetingId, Long userId);
+  Optional<MeetingMember> findByMeetingIdAndUserId(Integer meetingId, Integer userId);
 
-    boolean existsByMeetingIdAndUserId(Long meetingId, Long userId);
+  boolean existsByMeetingIdAndUserId(Integer meetingId, Integer userId);
 
-    boolean existsByMeetingIdAndUserIdAndRole(Long meetingId, Long userId, MeetingMemberRole role);
+  boolean existsByMeetingIdAndUserIdAndRole(Integer meetingId, Integer userId, MeetingMemberRole role);
 
-    List<MeetingMember> findByMeetingIdInAndRole(List<Long> meetingIds, MeetingMemberRole role);
+  List<MeetingMember> findByMeetingIdInAndRole(List<Integer> meetingIds, MeetingMemberRole role);
 
-    List<MeetingMember> findByMeetingIdIn(List<Long> meetingIds);
+  List<MeetingMember> findByMeetingIdIn(List<Integer> meetingIds);
 
-    List<MeetingMember> findByUserId(Long userId);
+  List<MeetingMember> findByUserId(Integer userId);
 }
