+++ backend/springboot/src/main/java/com/workmaite/domain/reports/dto/ReportReviewResultResponse.java
@@ -10,17 +10,17 @@ public class ReportReviewResultResponse {
 
     private Long reportId;
     private String aiStatus;
-    private Float totalScore;
+    private Double totalScore;
     private String detailScores;
     private String feedback;
 
-    public static ReportReviewResultResponse of(ReportScore reportScore) {
-        return ReportReviewResultResponse.builder()
-                .reportId(reportScore.getReportId())
-                .aiStatus(reportScore.getAiStatus())
-                .totalScore(reportScore.getTotalScore())
-                .detailScores(reportScore.getDetailScores())
-                .feedback(reportScore.getFeedback())
-                .build();
-    }
+  public static ReportReviewResultResponse of(ReportScore reportScore) {
+    return ReportReviewResultResponse.builder()
+        .reportId(reportScore.getReportId())
+        .aiStatus(reportScore.getAiStatus())
+        .totalScore(reportScore.getTotalScore())
+        .detailScores(reportScore.getDetailScores())
+        .feedback(reportScore.getFeedback())
+        .build();
+  }
 }
