+++ backend/springboot/src/main/java/com/workmaite/domain/reports/repository/ReportScoreRepository.java
@@ -1,10 +1,9 @@
 package com.workmaite.domain.reports.repository;
 
 import com.workmaite.domain.reports.entity.ReportScore;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface ReportScoreRepository extends JpaRepository<ReportScore, Long> {
-    Optional<ReportScore> findByReportId(Long reportId);
+public interface ReportScoreRepository extends JpaRepository<ReportScore, Integer> {
+  Optional<ReportScore> findByReportId(Integer reportId);
 }
