+++ backend/springboot/src/main/java/com/workmaite/domain/reports/service/ReportService.java
@@ -9,105 +9,109 @@ import com.workmaite.global.audit.AuditLogged;
 import com.workmaite.global.auth.MeetingAccessGuard;
 import com.workmaite.global.exception.BusinessException;
 import com.workmaite.global.exception.ErrorCode;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class ReportService {
 
-    private final ReportRepository reportRepository;
-    private final ReportScoreRepository reportScoreRepository;
-    private final MeetingAccessGuard meetingAccessGuard;
+  private final ReportRepository reportRepository;
+  private final ReportScoreRepository reportScoreRepository;
+  private final MeetingAccessGuard meetingAccessGuard;
 
-    @Transactional
-    @AuditLogged(action = "CREATE", entityType = "report")
-    public ReportResponse submitReport(Long meetingId, ReportSubmitRequest request) {
-        meetingAccessGuard.requireMember(meetingId);
-        Report report = Report.builder()
-                .meetingId(meetingId)
-                .uploadId(request.getUploadId())
-                .submitterDepartment(request.getSubmitterDepartment())
-                .fileName(request.getFileName())
-                .filePath(request.getFilePath())
-                .build();
-        return ReportResponse.of(reportRepository.save(report));
-    }
+  @Transactional
+  @AuditLogged(action = "CREATE", entityType = "report")
+  public ReportResponse submitReport(Integer meetingId, ReportSubmitRequest request) {
+    meetingAccessGuard.requireMember(meetingId);
+    Report report =
+        Report.builder()
+            .meetingId(meetingId)
+            .uploadId(request.getUploadId())
+            .submitterDepartment(request.getSubmitterDepartment())
+            .fileName(request.getFileName())
+            .filePath(request.getFilePath())
+            .build();
+    return ReportResponse.of(reportRepository.save(report));
+  }
 
-    @Transactional
-    @AuditLogged(action = "CREATE", entityType = "report")
-    public ReportResponse submitReportForSession(Long sessionId, ReportSessionSubmitRequest request) {
-        meetingAccessGuard.requireMember(request.getMeetingId());
-        Report report = Report.builder()
-                .meetingId(request.getMeetingId())
-                .uploadId(request.getUploadId())
-                .submitterDepartment(request.getSubmitterDepartment())
-                .fileName(request.getFileName())
-                .filePath(request.getFilePath())
-                .build();
-        return ReportResponse.of(reportRepository.save(report));
-    }
+  @Transactional
+  @AuditLogged(action = "CREATE", entityType = "report")
+  public ReportResponse submitReportForSession(Integer sessionId, ReportSessionSubmitRequest request) {
+    meetingAccessGuard.requireMember(request.getMeetingId());
+    Report report =
+        Report.builder()
+            .meetingId(request.getMeetingId())
+            .uploadId(request.getUploadId())
+            .submitterDepartment(request.getSubmitterDepartment())
+            .fileName(request.getFileName())
+            .filePath(request.getFilePath())
+            .build();
+    return ReportResponse.of(reportRepository.save(report));
+  }
 
-    @Transactional
-    @AuditLogged(action = "RESUBMIT", entityType = "report")
-    public ReportResponse resubmitReport(Long reportId, ReportResubmitRequest request) {
-        Report original = findByIdOrThrow(reportId);
+  @Transactional
+  @AuditLogged(action = "RESUBMIT", entityType = "report")
+  public ReportResponse resubmitReport(Integer reportId, ReportResubmitRequest request) {
+    Report original = findByIdOrThrow(reportId);
 
-        Report resubmitted = Report.builder()
-                .meetingId(original.getMeetingId())
-                .uploadId(request.getUploadId())
-                .submitterDepartment(request.getSubmitterDepartment())
-                .parentId(original.getId())
-                .version(original.getVersion() + 1)
-                .fileName(request.getFileName())
-                .filePath(request.getFilePath())
-                .build();
+    Report resubmitted =
+        Report.builder()
+            .meetingId(original.getMeetingId())
+            .uploadId(request.getUploadId())
+            .submitterDepartment(request.getSubmitterDepartment())
+            .parentId(original.getId())
+            .version(original.getVersion() + 1)
+            .fileName(request.getFileName())
+            .filePath(request.getFilePath())
+            .build();
 
-        return ReportResponse.of(reportRepository.save(resubmitted));
-    }
+    return ReportResponse.of(reportRepository.save(resubmitted));
+  }
 
-    public List<ReportResponse> getReportsByMeeting(Long meetingId) {
-        meetingAccessGuard.requireMember(meetingId);
-        return reportRepository.findAllByMeetingId(meetingId).stream()
-                .map(ReportResponse::of)
-                .toList();
-    }
+  public List<ReportResponse> getReportsByMeeting(Integer meetingId) {
+    meetingAccessGuard.requireMember(meetingId);
+    return reportRepository.findAllByMeetingId(meetingId).stream().map(ReportResponse::of).toList();
+  }
 
-    public ReportResponse getReport(Long reportId) {
-        return ReportResponse.of(findByIdOrThrow(reportId));
-    }
+  public ReportResponse getReport(Integer reportId) {
+    return ReportResponse.of(findByIdOrThrow(reportId));
+  }
 
-    @Transactional
-    @AuditLogged(action = "REVIEW", entityType = "report")
-    public ReportResponse updateHumanStatus(Long reportId, ReportStatusUpdateRequest request) {
-        Report report = findByIdOrThrow(reportId);
-        report.updateHumanStatus(request.getStatus());
-        return ReportResponse.of(report);
-    }
+  @Transactional
+  @AuditLogged(action = "REVIEW", entityType = "report")
+  public ReportResponse updateHumanStatus(Integer reportId, ReportStatusUpdateRequest request) {
+    Report report = findByIdOrThrow(reportId);
+    report.updateHumanStatus(request.getStatus());
+    return ReportResponse.of(report);
+  }
 
-    public ReportReviewResultResponse getReviewResult(Long reportId) {
-        meetingAccessGuard.requireMemberByReport(reportId);
-        ReportScore reportScore = reportScoreRepository.findByReportId(reportId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.REPORT_NOT_FOUND));
-        return ReportReviewResultResponse.of(reportScore);
-    }
+  public ReportReviewResultResponse getReviewResult(Integer reportId) {
+    meetingAccessGuard.requireMemberByReport(reportId);
+    ReportScore reportScore =
+        reportScoreRepository
+            .findByReportId(reportId)
+            .orElseThrow(() -> new BusinessException(ErrorCode.REPORT_NOT_FOUND));
+    return ReportReviewResultResponse.of(reportScore);
+  }
 
-    @Transactional
-    @AuditLogged(action = "DELETE", entityType = "report")
-    public void deleteReport(Long reportId) {
-        meetingAccessGuard.requireMemberByReport(reportId);
-        reportRepository.deleteById(reportId);
-    }
+  @Transactional
+  @AuditLogged(action = "DELETE", entityType = "report")
+  public void deleteReport(Integer reportId) {
+    meetingAccessGuard.requireMemberByReport(reportId);
+    reportRepository.deleteById(reportId);
+  }
 
-    // 모든 reportId 경로의 단일 진입점 — 멤버십 검증 포함 (IDOR 차단, P1-4)
-    private Report findByIdOrThrow(Long reportId) {
-        Report report = reportRepository.findById(reportId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.REPORT_NOT_FOUND));
-        meetingAccessGuard.requireMember(report.getMeetingId());
-        return report;
-    }
+  // 모든 reportId 경로의 단일 진입점 — 멤버십 검증 포함 (IDOR 차단, P1-4)
+  private Report findByIdOrThrow(Integer reportId) {
+    Report report =
+        reportRepository
+            .findById(reportId)
+            .orElseThrow(() -> new BusinessException(ErrorCode.REPORT_NOT_FOUND));
+    meetingAccessGuard.requireMember(report.getMeetingId());
+    return report;
+  }
 }
